+++ src/app/Controllers/Incidents/ListController.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App\Controllers;
+namespace App\Controllers\Incidents;
 
 use App\Core\Template;
 
 
-class AdminController
+class ListController
 {
     public function handle(Template $template, $pdo)
     {
